pyproject.toml
[tool.poetry]
name = "FitLife"
version = "0.1.0"
description = ""
authors = ["VCBorges <vinicius.costa.borges98@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
python-ulid = "^2.2.0"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
dj-database-url = "^2.1.0"
validate-docbr = "^1.10.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.11"
pytest-django = "^4.7.0"
rich = "^13.7.0"
django-extensions = "^3.2.3"
pytest-mock = "^3.12.0"
pygraphviz = "^1.12"
pre-commit = "^3.6.2"
ipython = "^8.23.0"
factory-boy = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["test_*.py"]
addopts = [
    "-vv", 
    "-s",
    "--reuse-db",
]

[tool.ruff]
exclude = [
    ".venv",
    "node_modules"
]

# Same as Black.
line-length = 88

# Assume Python 3.8
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = [
    "E4",
    "E7",
    "E9",
    "F",
    "I",
    "DJ",
]
ignore = ["DJ001"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []


# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "single"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"


# [tool.ruff.isort]
# known-third-party = [
#     "dynaconf",
# ]

[tool.ruff.lint.isort.sections]
# Group all Django imports into a separate section.
"django" = ["django"]

[tool.ruff.lint.isort]
section-order = [
    "future",
    "standard-library",
    "django",
    "first-party",
    "local-folder",
    "third-party",
]
